+++ internal/app/repositories/user_mock.go
@@ -119,11 +119,11 @@ func (mr *MockUserRepositoryMockRecorder) FindUserDetailsById(ctx, id any) *gomo
 }
 
 // List mocks base method.
-func (m *MockUserRepository) List(ctx context.Context, limit, offset int32) ([]models.User, int, error) {
+func (m *MockUserRepository) List(ctx context.Context, limit, offset uint64) ([]models.User, uint64, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "List", ctx, limit, offset)
 	ret0, _ := ret[0].([]models.User)
-	ret1, _ := ret[1].(int)
+	ret1, _ := ret[1].(uint64)
 	ret2, _ := ret[2].(error)
 	return ret0, ret1, ret2
 }
